Extract font code.
