+++ src/main/java/com/colapietro/throwback/lwjgl/HelloWorld.java
@@ -39,57 +39,15 @@ import static org.lwjgl.system.MemoryUtil.*;
  */
 public class HelloWorld {
 
-    private static final int WIDTH = 800;
-    private static final int HEIGHT = 600;
-    private int ww = 800;
-    private int wh = 600;
-    private int lineOffset;
-    private final int lineCount = 1;
-    private float lineHeight;
-    private final int scale  = 0;
-    private boolean kerningEnabled = true;
-    private boolean lineBBEnabled;
-    protected String text;
-    private final STBTTFontinfo info;
-    private final int ascent;
-    private final int descent;
-    private final int lineGap;
+    private static int WIDTH = 800;
+    private static int HEIGHT = 600;
+    private static final int NUMBER_OF_SUPPORTED_GLFW_JOYSTICKS = GLFW_JOYSTICK_LAST + 1;
 
-    private final ByteBuffer ttf;
     private long window;
     private Set<Integer> controllers;
     private Map<Integer, Boolean> controllersAdded;
     private Callback debugProc;
-    private final int fontHeight = 24;
-    private static final int NUMBER_OF_SUPPORTED_GLFW_JOYSTICKS = GLFW_JOYSTICK_LAST + 1;
-    int BITMAP_W = 512;
-    int BITMAP_H = 512;
-
-
-    public HelloWorld() {
-        try {
-            ttf = ioResourceToByteBuffer("FiraSans-Regular.ttf", 160 * 1024);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        info = STBTTFontinfo.create();
-        if (!stbtt_InitFont(info, ttf)) {
-            throw new IllegalStateException("Failed to initialize font information.");
-        }
-        text = "Bar";
-        try (MemoryStack stack = stackPush()) {
-            IntBuffer pAscent  = stack.mallocInt(1);
-            IntBuffer pDescent = stack.mallocInt(1);
-            IntBuffer pLineGap = stack.mallocInt(1);
-
-            stbtt_GetFontVMetrics(info, pAscent, pDescent, pLineGap);
-
-            ascent = pAscent.get(0);
-            descent = pDescent.get(0);
-            lineGap = pLineGap.get(0);
-        }
-    }
+    private Font font;
 
     private void run() {
         System.out.println("Hello LWJGL " + Version.getVersion() + "!");
@@ -107,6 +65,7 @@ public class HelloWorld {
     }
 
     private void init() {
+        font = new Font();
         GLFWErrorCallback.createPrint(System.err).set();
         if ( !glfwInit() ) {
             throw new IllegalStateException("Unable to initialize GLFW");
@@ -124,6 +83,7 @@ public class HelloWorld {
 
         final String title = "Hello World!";
         this.window = glfwCreateWindow(WIDTH, HEIGHT, title, NULL, NULL);
+        this.font.setWindow(window);
         if ( window == NULL ) {
             throw new RuntimeException("Failed to create the GLFW window");
         }
@@ -193,7 +153,7 @@ public class HelloWorld {
         } else if (event == GLFW_DISCONNECTED) {
             removeController(jid);
         }
-        text = controllers.size() + " controllers connected";
+        font.setText(controllers.size() + " controllers connected");
     }
 
     private void addController(int jid) {
@@ -228,7 +188,7 @@ public class HelloWorld {
         clearColor(RGBA.BLACK);
 
 
-        STBTTBakedChar.Buffer cdata = init(BITMAP_W, BITMAP_H);
+        STBTTBakedChar.Buffer cdata = font.init(font.BITMAP_W, font.BITMAP_H);
 
 
         while ( !glfwWindowShouldClose(window) ) {
@@ -241,20 +201,7 @@ public class HelloWorld {
             // invoked during this call.
             glfwPollEvents();
 
-            float scaleFactor = 1.0f + getScale() * 0.25f;
-
-
-            glPushMatrix();
-            // Zoom
-            glScalef(scaleFactor, scaleFactor, 1f);
-            // Scroll
-            glTranslatef(4.0f, getFontHeight() * 0.5f + 4.0f - getLineOffset() * getFontHeight(), 0f);
-
-            renderText(cdata, BITMAP_W, BITMAP_H);
-
-            glPopMatrix();
-
-            glfwSwapBuffers(getWindow());
+            font.loopRenderText(cdata);
         }
         cdata.free();
     }
@@ -318,18 +265,18 @@ public class HelloWorld {
         final Xbox360ControllerButton controllerButton = Xbox360ControllerButton.valueOf(buttonIndex);
         if(controllerButton.equals(Xbox360ControllerButton.A)) {
             clearColor(RGBA.GREEN);
-            text = "GREEN";
+            font.setText("GREEN");
         } else if(controllerButton.equals(Xbox360ControllerButton.B)) {
             clearColor(RGBA.RED);
-            text = "RED";
+            font.setText("RED");
         } else if(controllerButton.equals(Xbox360ControllerButton.Y)) {
             clearColor(RGBA.YELLOW);
-            text = "YELLOW";
+            font.setText("YELLOW");
         } else if(controllerButton.equals(Xbox360ControllerButton.X)) {
             clearColor(RGBA.BLUE);
-            text = "BLUE";
+            font.setText("BLUE");
         } else if (controllerButton.equals(Xbox360ControllerButton.BACK)) {
-            text = "BYE";
+            font.setText("BYE");
             glfwSetWindowShouldClose(window, true);
         }
     }
@@ -358,189 +305,26 @@ public class HelloWorld {
 
 
     private void windowSizeChanged(long window, int width, int height) {
-        this.ww = width;
-        this.wh = height;
+        this.WIDTH = width;
+        this.HEIGHT = height;
 
         glMatrixMode(GL_PROJECTION);
         glLoadIdentity();
         glOrtho(0.0, width, height, 0.0, -1.0, 1.0);
         glMatrixMode(GL_MODELVIEW);
 
-        setLineOffset(round(lineOffset));
+        setLineOffset(round(font.lineOffset));
     }
 
     private void setLineOffset(int offset) {
-        lineOffset = max(0, min(offset, lineCount - (int)(wh / lineHeight)));
+        font.lineOffset = max(0, min(offset, font.lineCount - (int)(HEIGHT / font.lineHeight)));
     }
 
     private static void framebufferSizeChanged(long window, int width, int height) {
         glViewport(0, 0, width, height);
     }
 
-    private STBTTBakedChar.Buffer init(int BITMAP_W, int BITMAP_H) {
-        int                   texID = glGenTextures();
-        STBTTBakedChar.Buffer cdata = STBTTBakedChar.malloc(96);
-
-        ByteBuffer bitmap = BufferUtils.createByteBuffer(BITMAP_W * BITMAP_H);
-        stbtt_BakeFontBitmap(this.ttf, getFontHeight(), bitmap, BITMAP_W, BITMAP_H, 32, cdata);
-
-        glBindTexture(GL_TEXTURE_2D, texID);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_ALPHA, BITMAP_W, BITMAP_H, 0, GL_ALPHA, GL_UNSIGNED_BYTE, bitmap);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-
-        glClearColor(43f / 255f, 43f / 255f, 43f / 255f, 0f); // BG color
-        glColor3f(169f / 255f, 183f / 255f, 198f / 255f); // Text color
-
-        glEnable(GL_TEXTURE_2D);
-        glEnable(GL_BLEND);
-        glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-
-        return cdata;
-    }
-
-    public int getFontHeight() {
-        return fontHeight;
-    }
-
     public long getWindow() {
         return window;
-    }
-
-    public int getScale() {
-        return scale;
-    }
-
-    public int getLineOffset() {
-        return lineOffset;
-    }
-
-    private void renderText(STBTTBakedChar.Buffer cdata, int BITMAP_W, int BITMAP_H) {
-        float scale = stbtt_ScaleForPixelHeight(info, getFontHeight());
-
-        try (MemoryStack stack = stackPush()) {
-            IntBuffer pCodePoint = stack.mallocInt(1);
-
-            FloatBuffer x = stack.floats(0.0f);
-            FloatBuffer y = stack.floats(0.0f);
-
-            STBTTAlignedQuad q = STBTTAlignedQuad.mallocStack(stack);
-
-            int lineStart = 0;
-
-            int i  = 0;
-            int to = text.length();
-
-            glBegin(GL_QUADS);
-            while (i < to) {
-                i += getCP(text, to, i, pCodePoint);
-
-                int cp = pCodePoint.get(0);
-                if (cp == '\n') {
-                    if (isLineBBEnabled()) {
-                        glEnd();
-                        renderLineBB(lineStart, i - 1, y.get(0), scale);
-                        glBegin(GL_QUADS);
-                    }
-
-                    y.put(0, y.get(0) + (ascent - descent + lineGap) * scale);
-                    x.put(0, 0.0f);
-
-                    lineStart = i;
-                    continue;
-                } else if (cp < 32 || 128 <= cp) {
-                    continue;
-                }
-
-                stbtt_GetBakedQuad(cdata, BITMAP_W, BITMAP_H, cp - 32, x, y, q, true);
-                if (isKerningEnabled() && i < to) {
-                    getCP(text, to, i, pCodePoint);
-                    x.put(0, x.get(0) + stbtt_GetCodepointKernAdvance(info, cp, pCodePoint.get(0)) * scale);
-                }
-
-                glTexCoord2f(q.s0(), q.t0());
-                glVertex2f(q.x0(), q.y0());
-
-                glTexCoord2f(q.s1(), q.t0());
-                glVertex2f(q.x1(), q.y0());
-
-                glTexCoord2f(q.s1(), q.t1());
-                glVertex2f(q.x1(), q.y1());
-
-                glTexCoord2f(q.s0(), q.t1());
-                glVertex2f(q.x0(), q.y1());
-            }
-            glEnd();
-            if (isLineBBEnabled()) {
-                renderLineBB(lineStart, text.length(), y.get(0), scale);
-            }
-        }
-    }
-
-    private void renderLineBB(int from, int to, float y, float scale) {
-        glDisable(GL_TEXTURE_2D);
-        glPolygonMode(GL_FRONT, GL_LINE);
-        glColor3f(1.0f, 1.0f, 0.0f);
-
-        float width = getStringWidth(info, text, from, to, getFontHeight());
-        y -= descent * scale;
-
-        glBegin(GL_QUADS);
-        glVertex2f(0.0f, y);
-        glVertex2f(width, y);
-        glVertex2f(width, y - getFontHeight());
-        glVertex2f(0.0f, y - getFontHeight());
-        glEnd();
-
-        glEnable(GL_TEXTURE_2D);
-        glPolygonMode(GL_FRONT, GL_FILL);
-        glColor3f(169f / 255f, 183f / 255f, 198f / 255f); // Text color
-    }
-
-    private float getStringWidth(STBTTFontinfo info, String text, int from, int to, int fontHeight) {
-        int width = 0;
-
-        try (MemoryStack stack = stackPush()) {
-            IntBuffer pCodePoint       = stack.mallocInt(1);
-            IntBuffer pAdvancedWidth   = stack.mallocInt(1);
-            IntBuffer pLeftSideBearing = stack.mallocInt(1);
-
-            int i = from;
-            while (i < to) {
-                i += getCP(text, to, i, pCodePoint);
-                int cp = pCodePoint.get(0);
-
-                stbtt_GetCodepointHMetrics(info, cp, pAdvancedWidth, pLeftSideBearing);
-                width += pAdvancedWidth.get(0);
-
-                if (isKerningEnabled() && i < to) {
-                    getCP(text, to, i, pCodePoint);
-                    width += stbtt_GetCodepointKernAdvance(info, cp, pCodePoint.get(0));
-                }
-            }
-        }
-
-        return width * stbtt_ScaleForPixelHeight(info, fontHeight);
-    }
-
-    private static int getCP(String text, int to, int i, IntBuffer cpOut) {
-        char c1 = text.charAt(i);
-        if (Character.isHighSurrogate(c1) && i + 1 < to) {
-            char c2 = text.charAt(i + 1);
-            if (Character.isLowSurrogate(c2)) {
-                cpOut.put(0, Character.toCodePoint(c1, c2));
-                return 2;
-            }
-        }
-        cpOut.put(0, c1);
-        return 1;
-    }
-
-    public boolean isKerningEnabled() {
-        return kerningEnabled;
-    }
-
-    public boolean isLineBBEnabled() {
-        return lineBBEnabled;
     }
 }
